feat(router): enforce runtime policies end-to-end, native streaming, live security reload
Policy enforcement (Law 2/5):
- attach the policy compiler pass per request from the live PolicyRegistry
  snapshot; both production call sites previously passed None and chat
  downgraded every rule to an audit-only stub
- admin API validates effects fail-closed and stores full declarations
  (the effect was previously discarded at write time)
- new RouterError::PolicyDenied: deny rules surface as 403 citing the rule
- E2E test proves create-deny -> block -> delete -> recover cycle

Streaming:
- single-node graphs stream natively via ProviderRouter::chat_stream wrapped
  in MeteredStream: mid-stream budget breach terminates, client disconnect
  releases the reservation, actuals replace estimates in accounting
- orchestrated graphs keep re-chunked SSE fallback; x-fusion-stream-mode
  header distinguishes transports

Security operations:
- AuthHandle + AuthReloader: SIGHUP hot-applies key rotation/scopes
- RateLimitReloader hot-applies limiter settings
- key scopes: key = chat, key:operator adds /v1/operations/*,
  /v1/executions, /metrics; unknown scopes rejected at load
- rate-limit middleware fails closed on missing extension

De-phantoming (~10k LOC removed):
- deleted crates: fusion-worker, fusion-worker-protocol, fusion-api-public,
  fusion-api-internal (fake replay engine), fusion-infrastructure,
  fusion-placement (constant-score placement), fusion-security
- preserved: SecretManager -> src/security/secrets.rs,
  ExecutionLeaseManager -> fusion_scheduler::leases (+ TTL/epoch tests)
- deleted dead monolith mass: eri/abi/target stack, trigger engine, package
  loader, intent module, capability resolver, capability_executor,
  distributed scheduler shim, calibration/connector_metrics/unified_diagnostics,
  capability_catalog, workflow registry wiring, dead strategies map
- deleted fiction-validating suites (beta/perf/placement/scheduler);
  conformance member lists updated to reality

Operational fixes:
- planner failure -> retryable 503 (PlannerFailure), no panic
- /ready pings telemetry DB + checks providers (503 when unhealthy)
- evidence snapshot 30s TTL cache + hourly 7-day retention job
- server-wide TimeoutLayer (request_timeout_secs, default 300) and
  ConcurrencyLimitLayer bound to resources.max_concurrent
- WorkQueue honors configured max_concurrent_nodes (was hardcoded 16)
- release builds refuse to boot when a provider key is unresolvable
- removed prompt-content debug logging in zen provider

Type consolidation:
- ModelCatalog: single canonical definition in fusion-core, re-exported by
  fusion-types (previously two independent drift-prone copies)

+++ crates/fusion-compiler/src/content_hash.rs
@@ -25,7 +25,7 @@ pub fn compute_workflow_content_hash(ir: &WorkflowIR) -> u64 {
 
     mix_u64(&mut h, ir.plan_id.as_u128() as u64);
     mix_u64(&mut h, (ir.plan_id.as_u128() >> 64) as u64);
-    mix_u64(&mut h, ir.metadata.policy_version as u64);
+    mix_u64(&mut h, ir.metadata.policy_version);
     mix_u64(&mut h, ir.nodes.len() as u64);
 
     let mut nodes: Vec<_> = ir.nodes.iter().collect();
